FEAT: Split tx wrapper in header and source
