+++ cli/inputhandler/CMakeLists.txt
@@ -3,7 +3,7 @@ file(GLOB inputhandlers
 	options_esdt.h options_esdt.cpp
 	arghandler.h arghandler.cpp
 	wrappers/iwrapper.h
-	wrappers/transaction_input_wrapper.h
+	wrappers/transaction_input_wrapper.h wrappers/transaction_input_wrapper.cpp
 	)
 
 add_library(inputhandler ${inputhandlers})
